// src/api-client/types.gen.ts
export type BotRun = {
  /**
   * Bot ID
   */
  bot_id: number;
  /**
   * Bot run ID
   */
  id?: string;
  /**
   * Timestamp when the run was created
   */
  inserted_at?: string;
  /**
   * Params used for the Bot run
   */
  params: {
    [key: string]: string;
  };
  /**
   * Result of the Bot run
   */
  result?: string;
};

/**
 * A list of Bot runs
 */
export type BotRunList = {
  [key: string]: unknown;
};

/**
 * Parameters for triggering a Bot run
 */
export type BotRunParams = {
  /**
   * Optional key-value pairs of parameters for the Bot run
   */
  params?: {
    [key: string]: string;
  };
};

/**
 * Status of a Bot run including run details and logs
 */
export type BotRunStatus = {
  /**
   * List of logs for the Bot run
   */
  logs: Array<{
    /**
     * Log level
     */
    level: "info" | "debug";
    /**
     * Log message
     */
    msg: string;
    /**
     * Timestamp when the log was created
     */
    timestamp: Date;
  }>;
  run: BotRun;
};